refactor(rest): extract SigV4 signing into catalog/rest/sigv4.py
